pyproject.toml
[project]
name = "brandforge"
version = "0.1.0"
description = "AI-Powered Multi-Agent Marketing Platform"
requires-python = "==3.11.*"
dependencies = [
    "google-adk>=0.3.0",
    "google-adk[a2a]",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "google-cloud-firestore>=2.16.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-pubsub>=2.21.0",
    "google-cloud-secret-manager>=2.18.0",
    "google-cloud-logging>=3.9.0",
    "google-cloud-texttospeech>=2.16.0",
    "reportlab>=4.0",
    "ffmpeg-python>=0.2.0",
    "opencv-python-headless>=4.8.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "Pillow>=10.0",
    "icalendar>=5.0",
    "google-cloud-scheduler>=2.13.0",
    "google-cloud-bigquery>=3.14.0",
    "google-cloud-monitoring>=2.19.0",
    "playwright>=1.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-timeout>=2.2",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["brandforge*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "gcp: marks tests requiring live GCP services (deselect with '-m \"not gcp\"')",
    "llm: marks tests requiring live Gemini API calls (deselect with '-m \"not llm\"')",
]
